// packages/client/src/defs.d.ts
export declare type Action = typeof Actions[keyof typeof Actions] | string;
export declare type ActionsAlias = typeof ActionsAliases[keyof typeof ActionsAliases] | 'custom' | null;
export declare type ActionsAliasStr = keyof typeof ActionsAliases;
export declare type Context = {
    action: Action;
    alias: ActionsAlias;
    model: string | null;
};
// ...
export declare type QueryParams = {
    type: GraphQLType;
    operation: string;
    context: Context;
    fields: string[];
    paths: string[];
    args: any;
    prismaArgs: PrismaArgs;
    authorization: Authorization;
    identity: Identity;
    headers: any;
};
export declare type Authorization = typeof Authorizations[keyof typeof Authorizations] | null;
export declare type QueryBuilder = {
    prismaGet: (prismaArgs: PrismaArgs) => any;
    prismaList: (prismaArgs: PrismaArgs) => any;
    prismaCount: (prismaArgs: PrismaArgs) => any;
    prismaCreate: (prismaArgs: PrismaArgs) => any;
    prismaCreateMany: (prismaArgs: PrismaArgs) => any;
    prismaUpdate: (prismaArgs: PrismaArgs) => any;
    prismaUpdateMany: (prismaArgs: PrismaArgs) => any;
    prismaUpsert: (prismaArgs: PrismaArgs) => any;
    prismaDelete: (prismaArgs: PrismaArgs) => any;
    prismaDeleteMany: (prismaArgs: PrismaArgs) => any;
    prismaWhere: (prismaArgs: PrismaArgs, filter: any) => PrismaArgs;
};
// ...
export declare type PrismaArgs = {
    where?: any;
    data?: any;
    orderBy?: any;
    skip?: number;
    take?: number;
    skipDuplicates?: boolean;
    select?: any;
};
export declare type AppsyncEvent = {
    arguments: any;
    source: any;
    identity: Identity;
    request: any;
    info: {
        fieldName: string;
        parentTypeName: string;
        variables: any;
        selectionSetList: string[];
        selectionSetGraphQL: string;
    };
    prev: {
        result: any;
    };
    stash: any;
};
export declare type GraphQLType = 'Query' | 'Mutation' | 'Subscription';
export declare type API_KEY = null | {
    [key: string]: any;
};
export declare type AWS_LAMBDA = {
    resolverContext: any;
    [key: string]: any;
};
export declare type AWS_IAM = {
    accountId: string;
    cognitoIdentityPoolId: string;
    cognitoIdentityId: string;
    sourceIp: string[];
    username: string;
    userArn: string;
    cognitoIdentityAuthType: string;
    cognitoIdentityAuthProvider: string;
    [key: string]: any;
};
export declare type AMAZON_COGNITO_USER_POOLS = {
    sub: string;
    issuer: string;
    username: string;
    claims: any;
    sourceIp: string[];
    defaultAuthStrategy: string;
    groups: string[];
    [key: string]: any;
};
export declare type OPENID_CONNECT = {
    claims: {
        sub: string;
        aud: string;
        azp: string;
        iss: string;
        exp: number;
        iat: number;
        gty: string;
    };
    sourceIp: string[];
    issuer: string;
    sub: string;
    [key: string]: any;
};
export declare type Identity = API_KEY | AWS_LAMBDA | AWS_IAM | AMAZON_COGNITO_USER_POOLS | OPENID_CONNECT;
// ...
export declare enum Authorizations {
    API_KEY = "API_KEY",
    AWS_IAM = "AWS_IAM",
    AMAZON_COGNITO_USER_POOLS = "AMAZON_COGNITO_USER_POOLS",
    AWS_LAMBDA = "AWS_LAMBDA",
    OPENID_CONNECT = "OPENID_CONNECT"
}